# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Die Groot Ompad — Integrated Travel Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Trip {
  id: number;
  name: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  endDate?: string | null;
  fuelPrice15?: number;
  fuelPrice18?: number;
  fuelPrice20?: number;
  revision?: string;
  createdAt: string;
  updatedAt: string;
}

export interface TripInput {
  name: string;
  notes?: string;
  startDate?: string;
  endDate?: string;
  fuelPrice15?: number;
  fuelPrice18?: number;
  fuelPrice20?: number;
}

export interface TripUpdate {
  name?: string;
  notes?: string;
  startDate?: string;
  endDate?: string;
  fuelPrice15?: number;
  fuelPrice18?: number;
  fuelPrice20?: number;
  revision?: string;
}

export interface TripSummary {
  tripId: number;
  totalLegs: number;
  totalPlannedKm: number;
  totalActualKm: number;
  totalEstFuelCost18: number;
  totalActualFuelCost: number;
  avgConsumptionL100km: number;
  kmVariance: number;
  costVariance: number;
}

export interface Leg {
  id: number;
  tripId: number;
  sortOrder: number;
  fromPlace: string;
  toPlace: string;
  plannedKm?: number;
  /** @nullable */
  actualKm?: number | null;
  /** @nullable */
  actualLitres?: number | null;
  /** @nullable */
  actualPricePerLitre?: number | null;
  /** @nullable */
  fuelPrice?: number | null;
  /** @nullable */
  serviceStation?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface LegInput {
  fromPlace: string;
  toPlace: string;
  plannedKm?: number;
  actualKm?: number;
  actualLitres?: number;
  actualPricePerLitre?: number;
  fuelPrice?: number;
  serviceStation?: string;
  notes?: string;
  sortOrder?: number;
}

export interface LegUpdate {
  fromPlace?: string;
  toPlace?: string;
  plannedKm?: number;
  actualKm?: number;
  actualLitres?: number;
  actualPricePerLitre?: number;
  fuelPrice?: number;
  serviceStation?: string;
  notes?: string;
  sortOrder?: number;
}

export interface LegReorder {
  legIds: number[];
}

export interface MonthBudget {
  fuel?: number;
  accommodation?: number;
  groceries?: number;
  eatingOut?: number;
  entertainment?: number;
  passesPermits?: number;
  insurance?: number;
  phoneInternet?: number;
  medical?: number;
  repairs?: number;
  subscriptions?: number;
  otherExpenses?: number;
  salary?: number;
  businessIncome?: number;
  refunds?: number;
  otherIncome1?: number;
  otherIncome2?: number;
  openingBalance?: number;
}

/**
 * Map of month index (0-11) to MonthBudget
 */
export type BudgetPlanMonths = {[key: string]: MonthBudget};

export interface BudgetPlan {
  id: number;
  tripId: number;
  year: string;
  /** Map of month index (0-11) to MonthBudget */
  months: BudgetPlanMonths;
  updatedAt: string;
}

export type BudgetPlanInputMonths = {[key: string]: MonthBudget};

export interface BudgetPlanInput {
  year: string;
  months: BudgetPlanInputMonths;
}

export interface MonthCashflow {
  month: number;
  income: number;
  expenses: number;
  closing: number;
}

export interface BudgetSummary {
  tripId: number;
  totalExpenses: number;
  totalIncome: number;
  netCashflow: number;
  monthlyBreakdown: MonthCashflow[];
}

export interface JournalEntry {
  id: number;
  tripId: number;
  weekDate: string;
  /** @nullable */
  whereWere?: string | null;
  /** @nullable */
  destinations?: string | null;
  /** @nullable */
  weather?: string | null;
  /** @nullable */
  weekSummary?: string | null;
  /** @nullable */
  loved?: string | null;
  /** @nullable */
  learned?: string | null;
  revision?: string;
  createdAt: string;
  updatedAt: string;
}

export interface JournalEntryInput {
  weekDate: string;
  whereWere?: string;
  destinations?: string;
  weather?: string;
  weekSummary?: string;
  loved?: string;
  learned?: string;
}

export interface JournalEntryUpdate {
  weekDate?: string;
  whereWere?: string;
  destinations?: string;
  weather?: string;
  weekSummary?: string;
  loved?: string;
  learned?: string;
}

export interface VehicleProfile {
  id: number;
  tripId: number;
  vehicleModel?: string;
  vehicleFuel?: string;
  kerbWeight?: number;
  gvm?: number;
  gcm?: number;
  towRating?: number;
  frontAxleLimit?: number;
  rearAxleLimit?: number;
  caravanModel?: string;
  caravanType?: string;
  caravanTare?: number;
  caravanAtm?: number;
  caravanGtm?: number;
  ballWeight?: number;
  waterLoad?: number;
  extrasLoad?: number;
  payloadPeople?: number;
  payloadFood?: number;
  payloadRecovery?: number;
  payloadTools?: number;
  payloadFuel?: number;
  payloadOther?: number;
  updatedAt: string;
}

export interface VehicleProfileInput {
  vehicleModel?: string;
  vehicleFuel?: string;
  kerbWeight?: number;
  gvm?: number;
  gcm?: number;
  towRating?: number;
  frontAxleLimit?: number;
  rearAxleLimit?: number;
  caravanModel?: string;
  caravanType?: string;
  caravanTare?: number;
  caravanAtm?: number;
  caravanGtm?: number;
  ballWeight?: number;
  waterLoad?: number;
  extrasLoad?: number;
  payloadPeople?: number;
  payloadFood?: number;
  payloadRecovery?: number;
  payloadTools?: number;
  payloadFuel?: number;
  payloadOther?: number;
}

export interface GpsPoint {
  id: number;
  tripId: number;
  lat: number;
  lng: number;
  /** @nullable */
  accuracy?: number | null;
  /** @nullable */
  note?: string | null;
  capturedAt: string;
}

export interface GpsPointInput {
  lat: number;
  lng: number;
  accuracy?: number;
  note?: string;
  capturedAt?: string;
}

export interface DashboardStats {
  totalTrips: number;
  totalKm: number;
  totalFuelCost: number;
  totalJournalEntries: number;
  recentTrips: Trip[];
}