# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Die Groot Ompad — Integrated Travel Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trips/{tripId}/legs:
    get:
      operationId: listLegs
      tags: [legs]
      summary: List all legs for a trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of legs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Leg"
    post:
      operationId: createLeg
      tags: [legs]
      summary: Add a leg to a trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LegInput"
      responses:
        "201":
          description: Leg created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Leg"
  /trips/{tripId}/legs/{legId}:
    patch:
      operationId: updateLeg
      tags: [legs]
      summary: Update a leg
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: legId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LegUpdate"
      responses:
        "200":
          description: Updated leg
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Leg"
    delete:
      operationId: deleteLeg
      tags: [legs]
      summary: Delete a leg
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: legId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /trips/{tripId}/legs/reorder:
    post:
      operationId: reorderLegs
      tags: [legs]
      summary: Reorder legs within a trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LegReorder"
      responses:
        "200":
          description: Reordered legs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Leg"

  # ── Budget ─────────────────────────────────────────────────────────────────
  /trips/{tripId}/budget:
    get:
      operationId: getBudget
      tags: [budget]
      summary: Get the 12-month budget plan for a trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Budget plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetPlan"
    put:
      operationId: saveBudget
      tags: [budget]
      summary: Save/replace the budget plan for a trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BudgetPlanInput"
      responses:
        "200":
          description: Saved budget plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetPlan"
  /trips/{tripId}/budget/summary:
    get:
      operationId: getBudgetSummary
      tags: [analytics]
      summary: Get yearly budget totals and cashflow by month
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Budget summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetSummary"

  # ── Journal ────────────────────────────────────────────────────────────────
  /trips/{tripId}/journal:
    get:
      operationId: listJournalEntries
      tags: [journal]
      summary: List all journal entries for a trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Journal entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JournalEntry"
    post:
      operationId: createJournalEntry
      tags: [journal]
      summary: Create a journal entry
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JournalEntryInput"
      responses:
        "201":
          description: Journal entry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"
  /trips/{tripId}/journal/{entryId}:
    patch:
      operationId: updateJournalEntry
      tags: [journal]
      summary: Update a journal entry
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: entryId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JournalEntryUpdate"
      responses:
        "200":
          description: Updated entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"
    delete:
      operationId: deleteJournalEntry
      tags: [journal]
      summary: Delete a journal entry
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: entryId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Vehicle ────────────────────────────────────────────────────────────────
...
  /trips/{tripId}/gps:
    get:
      operationId: listGpsPoints
      tags: [gps]
      summary: Get GPS track points for a trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: GPS points
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GpsPoint"
    post:
      operationId: logGpsPoint
      tags: [gps]
      summary: Log a GPS point
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GpsPointInput"
      responses:
        "201":
          description: GPS point logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GpsPoint"
    delete:
      operationId: clearGpsTrack
      tags: [gps]
      summary: Clear all GPS points for a trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Cleared

  # ── Analytics ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Leg:
      type: object
      required: [id, tripId, sortOrder, fromPlace, toPlace]
      properties:
        id:
          type: integer
        tripId:
          type: integer
        sortOrder:
          type: integer
        fromPlace:
          type: string
        toPlace:
          type: string
        plannedKm:
          type: number
        actualKm:
          type: ["number", "null"]
        actualLitres:
          type: ["number", "null"]
        actualPricePerLitre:
          type: ["number", "null"]
        fuelPrice:
          type: ["number", "null"]
        serviceStation:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    LegInput:
      type: object
      required: [fromPlace, toPlace]
      properties:
        fromPlace:
          type: string
        toPlace:
          type: string
        plannedKm:
          type: number
        actualKm:
          type: number
        actualLitres:
          type: number
        actualPricePerLitre:
          type: number
        fuelPrice:
          type: number
        serviceStation:
          type: string
        notes:
          type: string
        sortOrder:
          type: integer
    LegUpdate:
      type: object
      properties:
        fromPlace:
          type: string
        toPlace:
          type: string
        plannedKm:
          type: number
        actualKm:
          type: number
        actualLitres:
          type: number
        actualPricePerLitre:
          type: number
        fuelPrice:
          type: number
        serviceStation:
          type: string
        notes:
          type: string
        sortOrder:
          type: integer
    LegReorder:
      type: object
      required: [legIds]
      properties:
        legIds:
          type: array
          items:
            type: integer

    # ── Budget ─────────────────────────────────────────────────────────────────
    BudgetPlan:
      type: object
      required: [id, tripId, year, months, updatedAt]
      properties:
        id:
          type: integer
        tripId:
          type: integer
        year:
          type: string
        months:
          type: object
          description: Map of month index (0-11) to MonthBudget
          additionalProperties:
            $ref: "#/components/schemas/MonthBudget"
        updatedAt:
          type: string
    MonthBudget:
      type: object
      properties:
        fuel:
          type: number
        accommodation:
          type: number
        groceries:
          type: number
        eatingOut:
          type: number
        entertainment:
          type: number
        passesPermits:
          type: number
        insurance:
          type: number
        phoneInternet:
          type: number
        medical:
          type: number
        repairs:
          type: number
        subscriptions:
          type: number
        otherExpenses:
          type: number
        salary:
          type: number
        businessIncome:
          type: number
        refunds:
          type: number
        otherIncome1:
          type: number
        otherIncome2:
          type: number
        openingBalance:
          type: number
    BudgetPlanInput:
      type: object
      required: [year, months]
      properties:
        year:
          type: string
        months:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/MonthBudget"
    BudgetSummary:
      type: object
      required: [tripId, totalExpenses, totalIncome, netCashflow, monthlyBreakdown]
      properties:
        tripId:
          type: integer
        totalExpenses:
          type: number
        totalIncome:
          type: number
        netCashflow:
          type: number
        monthlyBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/MonthCashflow"
    MonthCashflow:
      type: object
      required: [month, income, expenses, closing]
      properties:
        month:
          type: integer
        income:
          type: number
        expenses:
          type: number
        closing:
          type: number

    # ── Journal ────────────────────────────────────────────────────────────────
    JournalEntry:
      type: object
      required: [id, tripId, weekDate, createdAt, updatedAt]
      properties:
        id:
          type: integer
        tripId:
          type: integer
        weekDate:
          type: string
        whereWere:
          type: ["string", "null"]
        destinations:
          type: ["string", "null"]
        weather:
          type: ["string", "null"]
        weekSummary:
          type: ["string", "null"]
        loved:
          type: ["string", "null"]
        learned:
          type: ["string", "null"]
        revision:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
    JournalEntryInput:
      type: object
      required: [weekDate]
      properties:
        weekDate:
          type: string
        whereWere:
          type: string
        destinations:
          type: string
        weather:
          type: string
        weekSummary:
          type: string
        loved:
          type: string
        learned:
          type: string
    JournalEntryUpdate:
      type: object
      properties:
        weekDate:
          type: string
        whereWere:
          type: string
        destinations:
          type: string
        weather:
          type: string
        weekSummary:
          type: string
        loved:
          type: string
        learned:
          type: string

    # ── Vehicle ────────────────────────────────────────────────────────────────
...
    GpsPoint:
      type: object
      required: [id, tripId, lat, lng, capturedAt]
      properties:
        id:
          type: integer
        tripId:
          type: integer
        lat:
          type: number
        lng:
          type: number
        accuracy:
          type: ["number", "null"]
        note:
          type: ["string", "null"]
        capturedAt:
          type: string
    GpsPointInput:
      type: object
      required: [lat, lng]
      properties:
        lat:
          type: number
        lng:
          type: number
        accuracy:
          type: number
        note:
          type: string
        capturedAt:
          type: string

    # ── Analytics ──────────────────────────────────────────────────────────────
...