# lib/api-zod/src/generated/types/gpsPoint.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Die Groot Ompad — Integrated Travel Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface GpsPoint {
  id: number;
  tripId: number;
  lat: number;
  lng: number;
  /** @nullable */
  accuracy?: number | null;
  /** @nullable */
  note?: string | null;
  capturedAt: string;
}