# lib/api-zod/src/generated/types/gpsPointInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Die Groot Ompad — Integrated Travel Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface GpsPointInput {
  lat: number;
  lng: number;
  accuracy?: number;
  note?: string;
  capturedAt?: string;
}