# lib/api-zod/src/generated/types/journalEntryUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Die Groot Ompad — Integrated Travel Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface JournalEntryUpdate {
  weekDate?: string;
  whereWere?: string;
  destinations?: string;
  weather?: string;
  weekSummary?: string;
  loved?: string;
  learned?: string;
}